pyproject.toml
[tool.poetry]
name = "ai01"
version = "0.2.2"
description = ""
authors = ["Om Gupta <omgupta0720@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
pydantic = "^1.8.1"
ruff = "0.6.9"
asyncio = "^3.4.3"
pyee = "11.1.0"
av = "11.0.0"
numpy = "^2.1.3"
websockets = "^13.1"
huddle01 = "1.0.4"
uuid = "^1.30"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"